feature(IDiscount.cs): The discount amount is now an integer.

+++ ObjectOrientedPractics/Model/IDiscount.cs
@@ -10,14 +10,14 @@
         /// </summary>
         /// <param name="items">Список товаров.</param>
         /// <returns>Размер скидки.</returns>
-        double Calculate(List<Item> items);
+        int Calculate(List<Item> items);
 
         /// <summary>
         /// Метод, который применяет скидку к товарам.
         /// </summary>
         /// <param name="items">Список товаров.</param>
         /// <returns>Размер скидки с вычетом накопленных баллов.</returns>
-        double Apply(List<Item> items);
+        int Apply(List<Item> items);
 
         /// <summary>
         /// Метод, который накапливает скидку или списывает баллы.
